Remove test_vrptw.py
